+++ apps/demo/app/page.tsx
@@ -2,5 +2,11 @@ import { RouteLoader } from '@vyuh/react-extension-content';
 
 export default function Home() {
   // Load the homepage route from CMS
-  return <RouteLoader url="/chakra" allowRefresh={true} />;
+  return (
+    <RouteLoader
+      url="/chakra"
+      allowRefresh={true}
+      live={process.env.NODE_ENV === 'development'}
+    />
+  );
 }