+++ packages/react-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vyuh/react-core
 
+## 0.2.3
+
 ## 0.2.2
 
 ### Patch Changes