+++ packages/react-extension-content/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vyuh/react-extension-content
 
+## 0.2.3
+
+### Patch Changes
+
+- @vyuh/react-core@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes